publish article

+++ src/main/java/com/wle/blog/vo/params/CommentParams.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CommentParams {
 
-    private Long articleId;
+    private Integer articleId;
 
     private String content;
